+++ HPXMLtoOpenStudio/resources/waterheater.rb
@@ -978,9 +978,10 @@ module Waterheater
 
     # Adjust COP based on RESNET HERS Addendum 77
     if not water_heating_system.hpwh_containment_volume.nil?
-      if water_heating_system.hpwh_installed_properly.nil? || (water_heating_system.hpwh_installed_properly == true)
-        runner.registerWarning("HPWHContainmentVolume provided in #{water_heating_system.id} is ignored because it is only used when HPWHInstalledProperly is 'false'.")
+      if water_heating_system.hpwh_confined_space_without_mitigation.nil? || (water_heating_system.hpwh_confined_space_without_mitigation == false)
+        runner.registerWarning("HPWHContainmentVolume provided in #{water_heating_system.id} is ignored because it is only used when HPWHInConfinedSpaceWithoutMitigation is 'true'.")
       else
+        # FUTURE: apply for 120V HPWH and other system types that the correction may not be accurate for
         if water_heating_system.hpwh_containment_volume < 450.0 && (water_heating_system.backup_heating_capacity == 0.0)
           runner.registerWarning("Heat pump water heater: #{water_heating_system.id} has no backup electric resistance element, COP adjustment for confined space may not be accurate when the containment space volume is below 450 cubic feet.")
         end